drivers: char: ipmi: Add Aspeed BMC SSIF driver
This driver implements the BMC side of SSIF interface for the Aspeed
AST25xx BMC.

It registers an I2C slave with address which is defined in ssif-bmc
node in device tree.  It also creates a device interface to user space and
supports system calls to handle IPMI request and response messages.

Signed-off-by: Thang Q. Nguyen <thang@os.amperecomputing.com>
Signed-off-by: Chuong Tran <chuong.tran@amperecomputing.com>