i2c:aspeed: workaround for slave’s "Unexpected Ack on read request"
Commit 2be6b47211e1 (“i2c: aspeed: Acknowledge most interrupts early in
interrupt handler”) moved interrupt acknowledgement to the beginning of
the interrupt handler. The side effect is that somehow it causes the
slave not to receive TX_ACK interrupt while in ASPEED_I2C_SLAVE_READ_PROCESSED
state. This results in “Unexpected Ack on read request” errors and
subsequently the driver cannot recover. This commit moves slave interrupt
acknowledgement to after all slave interrupts are processed as it used
to be before, and retain the logic for master interrupt acknowledgement.

Signed-off-by: Hieu Vu <hvu@amperecomputing.com>
Signed-off-by: Jim Saxman <jsaxman@amperecomputing.com>