hwmon:smpro: change the output of RAS internal APIs when no error
This commit fixes condition that buffer is undefined in case there is no error.
Initialize *buf to avoid undefined content in the case there is no error.
Also, eliminate the unecessary goto.

It also changes the return of errors_smpro and errors_pmpro APIs
when there is no error from output "0" to return empty.

Test:
	1. "cat errors_smpro" when there is no error`.
	2. The return should empty.

Signed-off-by: Thu B Nguyen <tbnguyen@amperecomputing.com>