misc: smpro-errmon: supports record error overflow events
Due to the changes of Host side, bit 8 of each RAS error counter register
will show the status of "error overflow" info. BMC will use this info
to record the SEL for this info. Be noted that, this bit will be cleared
when all error are handled (cleared) by BMC.
The Error Monitor driver will add one more error line in the output of sysfs
file when the error type is overflowed. The value of the Error Type and
Sub Type of that error line are "ff" and all characters of others fields
will be filled by 0.

Tested by:
	1. Issue a lot of RAS MCU error (current buffer for each error is 8)
	2. cat errors_mem_ce
	ff ff 0000 00000000 0000000000000000 0000000000000000\
		0000000000000000 0000000000000000 0000000000000000
	01 03 0000 41000000 0000000000000000 0000000000000000\
		0000000000000000 0000000000000000 0000000000000000
	01 04 0000 41000000 0000000000000000 0000000000000000\
		0000000000000000 0000000000000000 0000000000000000

Signed-off-by: Thu B Nguyen <tbnguyen@amperecomputing.com>